Drag Event Added and Button Crash Fixes
This commit adds the ability to scroll through the CodeEditor and stops the ObjectManager from causing crashes involving activing a deleted Button.

The Drag Event is far from finished, but it it is at least workable for now. Objects outside the bounds of the screen need to not be rendered. When an Object is added to the list off screen or the selecetd object goes off screen from adding an object, it needs to auto-focus to that object.