CommandObject Menu Implementation
Thanks to Chevron's work on the CommandObjects. I have implemented these Objects alongside the CodeEditor.

The CodeEditor draws each CommandObject in the right spot and still keeps drag functionality. The Draw of each CommandObject is changeable for each CommandObject.

The CommandLister now by default contains an array of String/Void() pairs. The String being the friendly display name and the Void() being the lambda expression which adds the command to the list of active commands.

I informed Chevron incorrectly of the size of the CommandObject. They should be 228*20.

The Old commandObject class is now listed as a commandListerItem and is responsible for properly listing which command will be added.